requirements.txt
requests
aiohttp
transformers
chromadb
tqdm
sentence-transformers
torch
openai
python-dotenv
fastapi
uvicorn
pydantic